move consts related to elemental used in shell to its own file in shell so that they can be used in the code without the Elemental package
