+++ pkg/elemental/types.js
@@ -2,7 +2,6 @@ export const ELEMENTAL_PRODUCT_NAME = 'elemental';
 export const ELEMENTAL_STORE = 'elemental';
 
 export const ELEMENTAL_PRODUCT_GROUP = 'elemental.cattle.io';
-export const ELEMENTAL_CLUSTER_PROVIDER = 'machineinventoryselectortemplate';
 export const ELEMENTAL_DEFAULT_NAMESPACE = 'fleet-default';
 
 export const ELEMENTAL_TYPES = {
@@ -22,15 +21,3 @@ export const ELEMENTAL_SCHEMAS = {
   MACHINE_INV_SELECTOR:           'elemental.cattle.io.machineinventoryselectors',
   MACHINE_INV_SELECTOR_TEMPLATES: 'elemental.cattle.io.machineinventoryselectortemplates',
 };
-
-export const ELEMENTAL_SCHEMA_IDS = {
-  MACHINE_INVENTORIES:            'elemental.cattle.io.machineinventory',
-  MACHINE_REGISTRATIONS:          'elemental.cattle.io.machineregistration',
-  MANAGED_OS_IMAGES:              'elemental.cattle.io.managedosimage',
-  MANAGED_OS_VERSION_CHANNELS:    'elemental.cattle.io.managedosversionchannel',
-  MANAGED_OS_VERSIONS:            'elemental.cattle.io.managedosversion',
-  MACHINE_INV_SELECTOR:           'elemental.cattle.io.machineinventoryselector',
-  MACHINE_INV_SELECTOR_TEMPLATES: 'elemental.cattle.io.machineinventoryselectortemplate',
-};
-
-export const KIND = { MACHINE_INV_SELECTOR_TEMPLATES: 'MachineInventorySelectorTemplate' };
